refactor: create report package for sending metrics to datadog #22
