+++ report/report.go
@@ -1,0 +1,2 @@
+// Package report provides sending metric to metrics services
+package report
